Update model tests for item
