+++ src/test/java/seedu/address/model/item/NameContainsKeywordsPredicateTest.java
@@ -1,4 +1,4 @@
-package seedu.address.newmodel.item;
+package seedu.address.model.item;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
