--- src/test/java/seedu/address/newmodel/ModelManagerTest.java
@@ -1,132 +1,0 @@
-package seedu.address.newmodel;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.newmodel.Model.PREDICATE_SHOW_ALL_ITEMS;
-import static seedu.address.testutil.Assert.assertThrows;
-import static seedu.address.testutil.TypicalItems.APPLE_PIE;
-import static seedu.address.testutil.TypicalItems.BANANA_MUFFIN;
-
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Arrays;
-
-import org.junit.jupiter.api.Test;
-
-import seedu.address.commons.core.GuiSettings;
-import seedu.address.newmodel.item.NameContainsKeywordsPredicate;
-import seedu.address.testutil.InventoryBuilder;
-
-public class ModelManagerTest {
-
-    private ModelManager modelManager = new ModelManager();
-
-    @Test
-    public void constructor() {
-        assertEquals(new UserPrefs(), modelManager.getUserPrefs());
-        assertEquals(new GuiSettings(), modelManager.getGuiSettings());
-        assertEquals(new Inventory(), new Inventory(modelManager.getInventory()));
-    }
-
-    @Test
-    public void setUserPrefs_nullUserPrefs_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> modelManager.setUserPrefs(null));
-    }
-
-    @Test
-    public void setUserPrefs_validUserPrefs_copiesUserPrefs() {
-        UserPrefs userPrefs = new UserPrefs();
-        userPrefs.setInventoryFilePath(Paths.get("inventory/file/path"));
-        userPrefs.setGuiSettings(new GuiSettings(1, 2, 3, 4));
-        modelManager.setUserPrefs(userPrefs);
-        assertEquals(userPrefs, modelManager.getUserPrefs());
-
-        // Modifying userPrefs should not modify modelManager's userPrefs
-        UserPrefs oldUserPrefs = new UserPrefs(userPrefs);
-        userPrefs.setInventoryFilePath(Paths.get("new/inventory/file/path"));
-        assertEquals(oldUserPrefs, modelManager.getUserPrefs());
-    }
-
-    @Test
-    public void setGuiSettings_nullGuiSettings_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> modelManager.setGuiSettings(null));
-    }
-
-    @Test
-    public void setGuiSettings_validGuiSettings_setsGuiSettings() {
-        GuiSettings guiSettings = new GuiSettings(1, 2, 3, 4);
-        modelManager.setGuiSettings(guiSettings);
-        assertEquals(guiSettings, modelManager.getGuiSettings());
-    }
-
-    @Test
-    public void setInventoryFilePath_nullPath_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> modelManager.setInventoryFilePath(null));
-    }
-
-    @Test
-    public void setInventoryFilePath_validPath_setsInventoryFilePath() {
-        Path path = Paths.get("inventory/file/path");
-        modelManager.setInventoryFilePath(path);
-        assertEquals(path, modelManager.getInventoryFilePath());
-    }
-
-    @Test
-    public void hasItem_nullItem_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> modelManager.hasItem(null));
-    }
-
-    @Test
-    public void hasItem_itemNotInInventory_returnsFalse() {
-        assertFalse(modelManager.hasItem(APPLE_PIE));
-    }
-
-    @Test
-    public void hasItem_itemInInventory_returnsTrue() {
-        modelManager.addItem(APPLE_PIE);
-        assertTrue(modelManager.hasItem(APPLE_PIE));
-    }
-
-    @Test
-    public void getFilteredItemList_modifyList_throwsUnsupportedOperationException() {
-        assertThrows(UnsupportedOperationException.class, () -> modelManager.getFilteredItemList().remove(0));
-    }
-
-    @Test
-    public void equals() {
-        Inventory inventory = new InventoryBuilder().withItem(APPLE_PIE).withItem(BANANA_MUFFIN).build();
-        Inventory differentInventory = new Inventory();
-        UserPrefs userPrefs = new UserPrefs();
-
-        // same values -> returns true
-        modelManager = new ModelManager(inventory, userPrefs);
-        ModelManager modelManagerCopy = new ModelManager(inventory, userPrefs);
-        assertTrue(modelManager.equals(modelManagerCopy));
-
-        // same object -> returns true
-        assertTrue(modelManager.equals(modelManager));
-
-        // null -> returns false
-        assertFalse(modelManager.equals(null));
-
-        // different types -> returns false
-        assertFalse(modelManager.equals(5));
-
-        // different inventory -> returns false
-        assertFalse(modelManager.equals(new ModelManager(differentInventory, userPrefs)));
-
-        // different filteredList -> returns false
-        String[] keywords = APPLE_PIE.getName().fullName.split("\\s+");
-        modelManager.updateFilteredItemList(new NameContainsKeywordsPredicate(Arrays.asList(keywords)));
-        assertFalse(modelManager.equals(new ModelManager(inventory, userPrefs)));
-
-        // resets modelManager to initial state for upcoming tests
-        modelManager.updateFilteredItemList(PREDICATE_SHOW_ALL_ITEMS);
-
-        // different userPrefs -> returns false
-        UserPrefs differentUserPrefs = new UserPrefs();
-        differentUserPrefs.setInventoryFilePath(Paths.get("differentFilePath"));
-        assertFalse(modelManager.equals(new ModelManager(inventory, differentUserPrefs)));
-    }
-}
