--- src/test/java/seedu/address/newmodel/UserPrefsTest.java
@@ -1,21 +1,0 @@
-package seedu.address.newmodel;
-
-import static seedu.address.testutil.Assert.assertThrows;
-
-import org.junit.jupiter.api.Test;
-
-public class UserPrefsTest {
-
-    @Test
-    public void setGuiSettings_nullGuiSettings_throwsNullPointerException() {
-        UserPrefs userPref = new UserPrefs();
-        assertThrows(NullPointerException.class, () -> userPref.setGuiSettings(null));
-    }
-
-    @Test
-    public void setInventoryFilePath_nullPath_throwsNullPointerException() {
-        UserPrefs userPrefs = new UserPrefs();
-        assertThrows(NullPointerException.class, () -> userPrefs.setInventoryFilePath(null));
-    }
-
-}
